Add files via upload
The base class HealthProfessional is a class used to represent a health professional and contains the attributes ID, name and basic information. It provides a default constructor and a constructor with parameters to initialize these properties. In addition, the class defines a series of Getter and Setter methods to ensure that these properties can be accessed and modified securely, as well as performing validity checks, such as ensuring that the ID is positive and that the name and basic information cannot be null. Finally, the printDetails() method is used to print the HealthProfessional's details, including ID, name, and basic information, making it easy for the user to view key information about the object. This design allows the HealthProfessional class to be well structured and extensible, and can be used as the basis for other specific HealthProfessional subclasses.