Add files via upload
The subclass GeneralPractitioner inherits from HealthProfessional and is specialized to represent general practitioners. The class contains three additional properties: department, specialty area and specialty details, which are used to describe the scope of the GP's work and expertise, respectively.

In terms of construction, GeneralPractitioner provides a default constructor that initializes all string properties to null, as well as a constructor with parameters that calls the constructor of the parent class to initialize the ID, name, and basic information, and to set up its own department, specialty area, and specialty details.

The class also contains a series of Getter and Setter methods to ensure secure access and modification of new properties. In these Setter methods, validity checks are performed to ensure that the Department, Specialty Area and Specialty Details cannot be null, thus maintaining data integrity.

In addition, GeneralPractitioner overrides the printDetails() method to first call the parent class method to print the basic information, then output the department, specialty area and specialty details, as well as specify the type of this health professional as a general practitioner. This design allows the GeneralPractitioner class to not only retain the characteristics of the base class, but also extend the specific information that applies to GPs, enhancing the extensibility and maintainability of the system.